+++ packages/destination-actions/src/destinations/stackadapt-audiences/deleteProfile/__tests__/onDelete.test.ts
@@ -1,6 +1,6 @@
 import nock from 'nock'
 import { createTestEvent, createTestIntegration } from '@segment/actions-core'
-import Definition from '../index'
+import Definition from '../../index'
 import { SegmentEvent } from '@segment/actions-core/*'
 
 const testDestination = createTestIntegration(Definition)
